services/agents/calculator_agent/requirements.txt
# Core dependencies
python-dotenv==1.0.0
python-dateutil==2.8.2
google-cloud-storage==2.10.0

# Calculator-specific dependencies
numexpr>=2.8.0  # Safe mathematical expression evaluation

# Note: vos_sdk is installed from local path via Dockerfile

# Metrics server
prometheus-client==0.19.0
flask==3.0.0
